+++ TinkoffASDKUI/TinkoffASDKUI/Modules/SBPPaymentSheet/Assembly/SBPPaymentSheetAssembly.swift
@@ -13,16 +13,16 @@ final class SBPPaymentSheetAssembly: ISBPPaymentSheetAssembly {
     // MARK: Dependencies
 
     private let acquiringSdk: AcquiringSdk
-    private let sbpConfiguration: SBPConfiguration
+    private let configuration: UISDKConfiguration
 
     // MARK: Initialization
 
     init(
         acquiringSdk: AcquiringSdk,
-        sbpConfiguration: SBPConfiguration
+        configuration: UISDKConfiguration
     ) {
         self.acquiringSdk = acquiringSdk
-        self.sbpConfiguration = sbpConfiguration
+        self.configuration = configuration
     }
 
     // MARK: ISBPPaymentSheetAssembly
@@ -35,7 +35,7 @@ final class SBPPaymentSheetAssembly: ISBPPaymentSheetAssembly {
             paymentStatusService: paymentStatusService,
             repeatedRequestHelper: repeatedRequestHelper,
             mainDispatchQueue: DispatchQueue.main,
-            sbpConfiguration: sbpConfiguration,
+            requestRepeatCount: configuration.paymentStatusRetriesCount,
             paymentId: paymentId
         )